Modified object as params handling in default route generation to trigger an error
